src/requirements.txt
# MemFlow Dependencies
# LlamaIndex core and readers
llama-index==0.10.68
llama-index-core==0.10.68
llama-index-embeddings-huggingface==0.2.3
llama-index-readers-file==0.1.33

# Local embedding model
sentence-transformers==2.7.0
transformers==4.44.2
torch==2.4.1

# File watching
watchdog==4.0.2

# MCP Server
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2

# Utilities
python-frontmatter==1.1.0
PyYAML==6.0.2

# Vector store
faiss-cpu==1.8.0.post1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
